Extract base analyzer class
This extracts common code from the analyzers into the `AnalyzerBase`
class.

The only requirements for an analyzer now are having the following
constants:

* `DEFAULT_PATHS` - default glob of files to analyze
* `LANGUAGE` - The analyzers language name as a string
* `DEFAULT_MASS_THRESHOLD` - Default mass we should tell Flay to use
* `BASE_POINTS` - Amount of points to multiply by s-expression mass

and the following method(s):

* `process_file(path)` - Returns s-expressions based on the file content
  of `path`
