+++ lib/cc/engine/analyzers/javascript/main.rb
@@ -1,3 +1,4 @@
+require "cc/engine/analyzers/analyzer_base"
 require "cc/engine/analyzers/javascript/parser"
 require "cc/engine/analyzers/javascript/node"
 require "cc/engine/analyzers/file_list"
@@ -8,37 +9,17 @@ module CC
   module Engine
     module Analyzers
       module Javascript
-        class Main
-          LANGUAGE = "javascript"
+        class Main < CC::Engine::Analyzers::Base
           DEFAULT_PATHS = [
             "**/*.js",
             "**/*.jsx"
           ]
+          LANGUAGE = "javascript"
           DEFAULT_MASS_THRESHOLD = 40
           BASE_POINTS = 3000
 
-          def initialize(engine_config:, directory:)
-            @engine_config = engine_config
-            @directory = directory
-          end
-
-          def run
-            files.map do |file|
-              process_file(file)
-            end
-          end
-
-          def mass_threshold
-            engine_config.mass_threshold_for(LANGUAGE) || DEFAULT_MASS_THRESHOLD
-          end
-
-          def base_points
-            BASE_POINTS
-          end
 
           private
-
-          attr_reader :engine_config, :directory
 
           def process_file(path)
             Node.new(js_parser.new(File.read(path), path).parse.syntax_tree, path).format
@@ -46,15 +27,6 @@ module CC
 
           def js_parser
             ::CC::Engine::Analyzers::Javascript::Parser
-          end
-
-          def files
-            ::CC::Engine::Analyzers::FileList.new(
-              directory: directory,
-              engine_config: engine_config,
-              default_paths: DEFAULT_PATHS,
-              language: LANGUAGE
-            ).files
           end
         end
       end
