+++ lib/cc/engine/analyzers/php/main.rb
@@ -1,4 +1,5 @@
 require 'cc/engine/analyzers/php/parser'
+require "cc/engine/analyzers/analyzer_base"
 require 'flay'
 require 'json'
 
@@ -6,7 +7,7 @@ module CC
   module Engine
     module Analyzers
       module Php
-        class Main
+        class Main < CC::Engine::Analyzers::Base
           LANGUAGE = "php"
           DEFAULT_PATHS = [
             "**/*.php",
@@ -16,30 +17,7 @@ module CC
           DEFAULT_MASS_THRESHOLD = 10
           BASE_POINTS = 4_000
 
-          attr_reader :directory, :engine_config, :io
-
-          def initialize(directory:, engine_config:)
-            @directory = directory
-            @engine_config = engine_config || {}
-          end
-
-          def run
-            files.map do |file|
-              process_file(file)
-            end
-          end
-
-          def mass_threshold
-            engine_config.mass_threshold_for(LANGUAGE) || DEFAULT_MASS_THRESHOLD
-          end
-
-          def base_points
-            BASE_POINTS
-          end
-
           private
-
-          attr_reader :engine_config, :directory
 
           def process_file(path)
             code = File.read(path)
@@ -51,19 +29,8 @@ module CC
             end
           end
 
-          private
-
           def php_parser
             ::CC::Engine::Analyzers::Php::Parser
-          end
-
-          def files
-            ::CC::Engine::Analyzers::FileList.new(
-              directory: directory,
-              engine_config: engine_config,
-              default_paths: DEFAULT_PATHS,
-              language: LANGUAGE
-            ).files
           end
         end
       end
